backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai>=1.40.0
sentry-sdk>=2.0.0
python-dotenv==1.0.0
